% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef_to_table}
\alias{coef_to_table}
\alias{coef_to_table_simple}
\title{Summarise model coefficients in a table}
\usage{
coef_to_table(fit, var_names = NULL, ..., min_coef = 1e-10)

coef_to_table_simple(fit, ..., min_coef = 1e-10)
}
\arguments{
\item{fit}{A fitted model with coefficients (e.g. from glmnet).}

\item{var_names}{(optional) A character vector of column names for the data used in \code{fit}.}

\item{...}{Parameters passed to \code{coef()}. For \code{cv.glmnet} models this will most usually be the
the regularisation parameter \code{s} of the desired model.}

\item{min_coef}{Coefficients with smaller absolute value than this are excluded from the table.}
}
\value{
A tibble which will always contain \code{name} and \code{coef} columns. If \code{var_names} is supplied
then there will be extra columns also. If there are no interactions then there will be two
extra columns \code{var} and \code{level} which are extracted from \code{name}. If there are interactions,
there will instead be three columns for each interaction: \code{interact_dd}, \code{var_dd}, and
\code{level_dd}, where each "dd" gives the number of the interaction. The number of interactions
displayed depends on which coefficients are included in the returned table, not the original
model, so \code{min_coef} can change the number of output columns.
}
\description{
Returns a table of coefficients, filtered by minimum absolute value and optionally separating
into interactions and decomposing dummy variables into parent variable and factor level parts.
}
\details{
Only coefficients with absolute value greater than \code{min_coef} are included. If \code{var_names} is
supplied then extra columns will be added which split the \code{name} into component parts. Where
there are interactions, these are separated, and any dummy variables are split into their
original variable and factor level names. See "Value" below for more details.

This was designed for glmnet models, but it also often works with other glm-type models.
}
\keyword{internal}
