% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{contrast_one_hot}
\alias{contrast_one_hot}
\title{Contrast function for one-hot encodings}
\usage{
contrast_one_hot(n, contrasts = TRUE, sparse = FALSE)
}
\arguments{
\item{n}{A vector of character factor levels or the number of unique levels.}

\item{contrasts}{Only the default of \code{TRUE} is supported.}

\item{sparse}{Only the default of \code{FALSE} is supported.}
}
\value{
A diagonal matrix that is \code{n}-by-\code{n}.
}
\description{
This contrast function produces a model matrix that has indicator columns for
each level of each factor.
}
\details{
Only intended to be used in \code{one_hot()}. Copy of \code{hardhat:::contr_one_hot()}.
}
\keyword{internal}
