% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{expand_interaction}
\alias{expand_interaction}
\title{Expand dummy names to var and level parts with appropriate column names}
\usage{
expand_interaction(tbl, var_names)
}
\arguments{
\item{tbl}{A tibble with one column called \code{name_dd} where "dd" are digits.}

\item{var_names}{A character vector of column names for the data used in \code{tbl}.}
}
\description{
Helper for \code{coef_to_table()}. This is used where there are interactions (which will be numbered).
Splits supplied name column into variable and level
parts. The returned tibble has columns \code{interact_dd} (input renamed),\code{var_dd} (variable name
part) and \code{level_dd} (factor level part), where "dd" matches
the value in the input column name \code{name_dd}.
}
\keyword{internal}
