% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{extract_level}
\alias{extract_level}
\title{Extract a factor level name from a model.matrix level names}
\usage{
extract_level(name_vec, feature_vec, none_name = NA_character_)
}
\arguments{
\item{name_vec}{Character vector of model.matrix names.}

\item{feature_vec}{Character vector of feature names corresponding to \code{name_vec}.}

\item{none_name}{String to use when there is no level or interaction in a model name.}
}
\description{
Helper for \code{coef_to_table()}. The model.matrix names are in \code{name_vec} and the
corresponding entries in \code{feature_vec} is the name of the first feature. A vector of levels
corresponding to the level associated with the feature is returned. If there is no level involved
then the level column will have the entry \code{none_name}.
}
\keyword{internal}
