% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_predict_binom.R
\name{glmnet_predict_binom}
\alias{glmnet_predict_binom}
\title{Consistent formatting for binomial cv.glmnet predictions}
\usage{
glmnet_predict_binom(fit, newx, type = "response", ...)
}
\arguments{
\item{fit}{A model fitted with \code{glmnet::cv.glmnet()}.}

\item{newx}{Matrix of new values for \code{x} at which predictions are to be made. See \code{glmnet::predict.glmnet()}.}

\item{type}{Passed to \code{predict()} with default "response". See \code{?glmnet::predict.glmnet}.}

\item{...}{Other arguments passed to \code{predict()}.}
}
\value{
A tibble of prediction probabilities with a column for each class.
}
\description{
Gets response predictions from a binomial \code{cv.glmnet} model in the same form as predictions from
ranger models and as used by yardstick.
}
