% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{glmnet_to_table2}
\alias{glmnet_to_table2}
\title{Summarise coefficients from glmnet in a table}
\usage{
glmnet_to_table2(
  fit,
  data = NULL,
  var_names = NULL,
  s = "lambda.1se",
  min_coef = 1e-10
)
}
\arguments{
\item{fit}{A fitted glmnet model.}

\item{data}{The data used to fit the model.}

\item{var_names}{(optional) A character vector of column names for data used in \code{fit}.}

\item{s}{The regularisation parameter determines which model is used from \code{fit} (as used in glmnet).}

\item{min_coef}{Coefficients with smaller absolute value than this are excluded from the table.}
}
\description{
Experimental variant of \code{glmnet_to_table()} which splits by interaction when finding parents.
}
\details{
This is still in development. Doesn't handle more than a single interaction.
}
