% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ilogit}
\alias{ilogit}
\title{Vectorised inverse logit function}
\usage{
ilogit(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
exp(x)/(1+exp(x))
}
\description{
Calculate the inverse logit function \eqn{exp(x) / (1 + exp(x))}. Modified from Faraway package.
I added a check on \code{x} to avoid \code{NaN} in the output, which would occur when \code{x} is greater than
about 750. Since \eqn{exp(x) / (1 + exp(x)) = 1} for \eqn{x>=20}, I only check for x > 20 and
output 1 for these cases.
}
