% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jemodel-package.R
\docType{package}
\name{jemodel-package}
\alias{jemodel}
\alias{jemodel-package}
\title{jemodel: Personal package to help with modelling}
\description{
Some functions to assist with modelling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jedwards24/jemodel}
  \item Report bugs at \url{https://github.com/jedwards24/jemodel/issues}
}

}
\author{
\strong{Maintainer}: James Edwards \email{jamese1326@gmail.com}

}
\keyword{internal}
