% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{match_parent}
\alias{match_parent}
\title{Match model level names to parent variable names}
\usage{
match_parent(level_names, parent_names)
}
\arguments{
\item{level_names}{Character vector to find parents for.}

\item{parent_names}{Character vector of candidate parents.}
}
\description{
Helper for \code{glmnet_to_table()}. Each level starts with a parent plus maybe more text. Returns
vector of parents corresponding to \code{level_names}. Parent is matching parent of greatest length.
If no match is found or level is \code{NA} then the parent is the level name.
}
\keyword{internal}
