% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{one_hot}
\alias{one_hot}
\title{Model matrix with one-hot encoding for unordered factors}
\usage{
one_hot(formula, data, intercept = FALSE)
}
\arguments{
\item{formula}{A model formula passed to \code{model.matrix()}.}

\item{data}{A data frame passed to \code{model.matrix()}.}

\item{intercept}{If \code{FALSE} (the default) then the first column of the matrix is
removed before returning/ If \code{TRUE} then the returned matrix is unaffected.}
}
\value{
A matrix (as from \code{model.matrix()}).
}
\description{
This replicates \code{model.matrix()} but with one-hot encoding
for unordered factors. Ordered factors use the default contrasts (as
set in \code{options("contrasts")}. By default, the first column of the model matrix
(the intercept) is removed.
}
\details{
The contrast function used is \code{contr_one_hot()}.
}
