% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{oob_errors}
\alias{oob_errors}
\title{Helper for \code{rang_oob_err()}}
\usage{
oob_errors(oob_mat, n_trees, target)
}
\arguments{
\item{oob_mat}{Matrix of out-of-bag individual tree class predicitions with an \code{NA} where the prediciton
was in bag. Rows are observation, columns are trees.}

\item{n_trees}{Positive integer. Gives the number of trees to be used in the prediction (columns \code{1:n_trees}
of \code{oob_mat} will be used).}

\item{target}{Vector of true classes (either 1 or 2).}
}
\description{
Returns vector of error indicators chosen by majority vote. Where the vote is split the entry wil be 0.5.
The entry will be \code{NaN} Where there are no out-of-bag predictions for a data point.
}
\keyword{internal}
