% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{rang_mtry}
\alias{rang_mtry}
\title{Tune ranger models for \code{mtry}}
\usage{
rang_mtry(
  data,
  fmla,
  m_vec,
  seed = 1,
  num.trees = 200,
  respect.unordered.factors = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame containing both input and target variables.}

\item{fmla}{A formula for the model.}

\item{m_vec}{Integer vector of values of tuning parameter \code{mtry}.}

\item{seed, num.trees, respect.unordered.factors}{Optional arguments passed to \code{ranger()}. Defaults are
\code{1}, \code{200} and \code{TRUE} respectively.}

\item{...}{Other arguments passed to \code{ranger}.}
}
\description{
Fits \code{ranger()} models for a given range of values of \code{mtry}. Output is a table and graph giving
OOB errors.
}
