% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{rang_roc_cut}
\alias{rang_roc_cut}
\title{ROC curve optimal cut for a ranger object}
\usage{
rang_roc_cut(rf, actual, class_name = NULL, plot = TRUE)
}
\arguments{
\item{rf}{A ranger fitted model.}

\item{actual}{A binary class target vector matching \code{rf}.}

\item{class_name}{String giving name of the class predicted in \code{actual}.}

\item{plot}{Produce a plot. Defaults to \code{TRUE}.}
}
\description{
Calls \code{roc_cut()} for a fitted ranger model. See \code{roc_cut()} for details. Requires
\code{probability = TRUE} to be used when the ranger model was fitted.
}
\details{
The \code{rf$predictions} is a matrix with columns for each class. The class corresponding to
the predictions in \code{actual} can be named in \code{class_name}. Default is to select the "maximum"
column name i.e. "TRUE" if classes are TRUE/FALSE and the largest number if numeric.
}
