% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardise}
\alias{standardise}
\alias{normalise}
\title{Standardise or normalise a numeric vector}
\usage{
standardise(x)

normalise(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
\itemize{
\item \code{standardise()} centres and scales \code{x} to \eqn{(x - mean(x)) / sd(x)}. Output has mean 0
and standard deviation 1.
\item \code{normalise()} normalises \code{x} to \eqn{(x - min(x)) / (max(x) - min(x))}. Output is
in [0,1].
}

Missing values are ignored in all calculations but are included in the returned vector.
}
