% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{extract_level}
\alias{extract_level}
\title{Extract a factor level name from a model.matrix level names}
\usage{
extract_level(name_vec, feature_vec)
}
\arguments{
\item{name_vec}{Character vector of model.matrix names.}

\item{feature_vec}{Character vector of feature names corresponding to \code{name_vec}.}
}
\description{
Helper for \code{glmnet_to_table()}. The model.matrix names are in \code{name_vec} and the
corresponding entries in \code{feature_vec} is the name of the first feature. A vector of levels
corresponding to the level associated with the feature is returned. If the is no level then the entry is
"(none)".
}
\keyword{internal}
